Review feedback: rename Roster tests
In this commit, the existing tests for RFC 6121 Section 2 ("Managing the Roster") are renamed to match the conventions used by pre-existing test implementations for section 8.
