+++ src/main/java/org/igniterealtime/smack/inttest/rfc6121/section2/RFC6121Section2_4_UpdateIntegrationTest.java
@@ -1,4 +1,4 @@
-package org.igniterealtime.smack.inttest.rfc6121;
+package org.igniterealtime.smack.inttest.rfc6121.section2;
 
 import org.igniterealtime.smack.inttest.AbstractSmackIntegrationTest;
 import org.igniterealtime.smack.inttest.SmackIntegrationTestEnvironment;
@@ -38,13 +38,13 @@ import static org.junit.jupiter.api.Assertions.*;
  * @author Guus der Kinderen, guus.der.kinderen@gmail.com
  */
 @SpecificationReference(document = "RFC6121")
-public class RFC6121Section2dot4RosterIntegrationTest extends AbstractSmackIntegrationTest
+public class RFC6121Section2_4_UpdateIntegrationTest extends AbstractSmackIntegrationTest
 {
     private final boolean isSendPresence;
 
     private final SmackIntegrationTestEnvironment environment;
 
-    public RFC6121Section2dot4RosterIntegrationTest(SmackIntegrationTestEnvironment environment) throws SmackException.NotConnectedException, SmackException.NoResponseException, InterruptedException, TestNotPossibleException
+    public RFC6121Section2_4_UpdateIntegrationTest(SmackIntegrationTestEnvironment environment) throws SmackException.NotConnectedException, SmackException.NoResponseException, InterruptedException, TestNotPossibleException
     {
         super(environment);
